make the bootstrap JS work again
The bundling in main.py had the bootstrap JS before
jquery, causing the bootstrap JS to not work.

Additionally, it now imports the JS from the bootstrap
dist directory, and uses the bootstrap.bundle.js from there
so we get the popper library included. This also removes the
bootstrap-rtl for now, as not sure where i got it originally.