coreos: add 'stable' option in streams dropdown
Allow switching between the 'stable' and 'testing' streams by
making a selection in the dropdown menu. Upon selection, a
refresh of the loaded `streamData` is triggered by modifying
the watched `stream` data member, and new metadata is fetched.

In order to use the Vue watch between the jumbotron buttons and
the rest of the app, state data (`data`) is moved back into a
single Vue instance, along with methods associated with the
jumbotron buttons.

The width of the jumbotron buttons menu is slightly altered in
this change by extending to the same width as the main view app,
rather than the full page.

This is a quick change in order to have the switching between
streams working - further improvements to the visual aspects of
the page may be made afterward.

Signed-off-by: Robert Fairley <rfairley@redhat.com>