coreos-download: fix navbar and stream dropdown
Fixes a bug that when the page re-renders, the navbar
and stream selection dropdown are not consistent with
the content shown in the main content section.

To reproduce, select a platform in the navbar and select a
different stream from dropdown. When the page re-renders,
while the main content remains the same as user selected
previously, the selected dropdown and navbar are refreshed
with default values. The expected behaviour is that the
refreshed value is consistent with the content selected by
user.

Signed-off-by: Allen Bai <abai@redhat.com>