coreos-release-note: fix error produced when calculating pkgdiffs
This error was not found previously since the conditional block was
not hit when calculating pkgdiffs. Essentially the accumulated pkgdiff
is an array of arrays, and the previous code lacks a layer of indexing
to modify the individual pkgdiff info.

In the future, more error checkings need to be added to handle this kind
of situations more gracefully.

Signed-off-by: Allen Bai <abai@redhat.com>